+++ telegram/client.py
@@ -11,6 +11,7 @@ import threading
 from typing import Any, Dict, List, Type, Callable, Optional, DefaultDict, Tuple, Union
 from types import FrameType
 from collections import defaultdict
+import enum
 
 from telegram import VERSION
 from telegram.utils import AsyncResult
@@ -21,6 +22,16 @@ logger = logging.getLogger(__name__)
 
 
 MESSAGE_HANDLER_TYPE: str = 'updateNewMessage'
+
+
+class AuthorizationState(enum.Enum):
+    NONE = None
+    WAIT_CODE = 'authorizationStateWaitCode'
+    WAIT_PASSWORD = 'authorizationStateWaitPassword'
+    WAIT_TDLIB_PARAMETERS = 'authorizationStateWaitTdlibParameters'
+    WAIT_ENCRYPTION_KEY = 'authorizationStateWaitEncryptionKey'
+    WAIT_PHONE_NUMBER = 'authorizationStateWaitPhoneNumber'
+    READY = 'authorizationStateReady'
 
 
 class Telegram:
@@ -80,6 +91,7 @@ class Telegram:
         self.proxy_port = proxy_port
         self.proxy_type = proxy_type
         self.use_secret_chats = use_secret_chats
+        self.authorization_state = AuthorizationState.NONE
 
         if not self.bot_token and not self.phone:
             raise ValueError('You must provide bot_token or phone')
@@ -463,45 +475,84 @@ class Telegram:
 
         return self._send_data(data, result_id='getAuthorizationState')
 
-    def login(self) -> None:
+    def _wait_authorization_result(self, result: AsyncResult) -> AuthorizationState:
+        authorization_state = None
+        if result:
+            result.wait(raise_exc=True)
+
+            if result.update is None:
+                raise RuntimeError('Something wrong, the result update is None')
+
+            if result.id == 'getAuthorizationState':
+                authorization_state = result.update['@type']
+            else:
+                authorization_state = result.update['authorization_state']['@type']
+
+        return AuthorizationState(authorization_state)
+
+    def login(self, blocking: bool = True) -> AuthorizationState:
         """
-        Login process (blocking)
+        Login process.
 
         Must be called before any other call.
         It sends initial params to the tdlib, sets database encryption key, etc.
+
+        args:
+          blocking [bool]: If True, the process is blocking and the client
+                           expects password and code from stdin.
+                           If False, `login` call returns next AuthorizationState and
+                           the login process can be continued (with calling login(blocking=False) again)
+                           after the necessary action is completed.
+
+        Returns:
+         - AuthorizationState.WAIT_CODE if a telegram code is required.
+           The caller should ask the telegram code
+           to the end user then call send_code(code)
+         - AuthorizationState.WAIT_PASSWORD if a telegram password is required.
+           The caller should ask the telegram password
+           to the end user and then call send_password(password)
+         - AuthorizationState.READY if the login process scceeded.
         """
         if self.proxy_server:
             self._send_add_proxy()
 
-        authorization_state = None
-        actions = {
-            None: self.get_authorization_state,
-            'authorizationStateWaitTdlibParameters': self._set_initial_params,
-            'authorizationStateWaitEncryptionKey': self._send_encryption_key,
-            'authorizationStateWaitPhoneNumber': self._send_phone_number_or_bot_token,
-            'authorizationStateWaitCode': self._send_telegram_code,
-            'authorizationStateWaitPassword': self._send_password,
-            'authorizationStateReady': self._complete_authorization,
+        actions: Dict[
+            AuthorizationState,
+            Callable[[], AsyncResult],
+        ] = {
+            AuthorizationState.NONE: self.get_authorization_state,
+            AuthorizationState.WAIT_TDLIB_PARAMETERS: self._set_initial_params,
+            AuthorizationState.WAIT_ENCRYPTION_KEY: self._send_encryption_key,
+            AuthorizationState.WAIT_PHONE_NUMBER: self._send_phone_number_or_bot_token,
+            AuthorizationState.WAIT_CODE: self._send_telegram_code,
+            AuthorizationState.WAIT_PASSWORD: self._send_password,
         }
+
+        blocking_actions = (
+            AuthorizationState.WAIT_CODE,
+            AuthorizationState.WAIT_PASSWORD,
+        )
+
         if self.phone:
             logger.info('[login] Login process has been started with phone')
         else:
             logger.info('[login] Login process has been started with bot token')
 
-        while not self._authorized:
-            logger.info('[login] current authorization state: %s', authorization_state)
-            result = actions[authorization_state]()
+        while self.authorization_state != AuthorizationState.READY:
+            logger.info(
+                '[login] current authorization state: %s', self.authorization_state
+            )
 
-            if result:
-                result.wait(raise_exc=True)
+            if not blocking and self.authorization_state in blocking_actions:
+                return self.authorization_state
 
-                if result.update is None:
-                    raise RuntimeError('Something wrong, the result update is None')
+            result = actions[self.authorization_state]()
+            if not isinstance(result, AuthorizationState):
+                self.authorization_state = self._wait_authorization_result(result)
+            else:
+                self.authorization_state = result
 
-                if result.id == 'getAuthorizationState':
-                    authorization_state = result.update['@type']
-                else:
-                    authorization_state = result.update['authorization_state']['@type']
+        return self.authorization_state
 
     def _set_initial_params(self) -> AsyncResult:
         logger.info(
@@ -580,20 +631,56 @@ class Telegram:
 
         return self._send_data(data, result_id='updateAuthorizationState')
 
-    def _send_telegram_code(self) -> AsyncResult:
+    def _send_telegram_code(self, code: Optional[str] = None) -> AsyncResult:
         logger.info('Sending code')
-        code = input('Enter code:')
+        if code is None:
+            code = input('Enter code:')
         data = {'@type': 'checkAuthenticationCode', 'code': str(code)}
 
         return self._send_data(data, result_id='updateAuthorizationState')
 
-    def _send_password(self) -> AsyncResult:
+    def send_code(self, code: str) -> AuthorizationState:
+        """
+        Verifies a telegram code and continues the authorization process
+
+        Args:
+          code: the code to be verified. If code is None, it will be asked to the user using the input() function
+
+        Returns
+         - AuthorizationState. The called have to call `login` to continue the login process.
+
+        Raises:
+         - RuntimeError if the login failed
+        """
+        result = self._send_telegram_code(code)
+        self.authorization_state = self._wait_authorization_result(result)
+
+        return self.authorization_state
+
+    def _send_password(self, password: Optional[str] = None) -> AsyncResult:
         logger.info('Sending password')
-        password = getpass.getpass('Password:')
+        if password is None:
+            password = getpass.getpass('Password:')
         data = {'@type': 'checkAuthenticationPassword', 'password': password}
 
         return self._send_data(data, result_id='updateAuthorizationState')
 
-    def _complete_authorization(self) -> None:
-        logger.info('Completing auth process')
-        self._authorized = True
+    def send_password(self, password: str) -> AuthorizationState:
+        """
+        Verifies a telegram password and continues the authorization process
+
+        Args:
+          password the password to be verified.
+          If password is None, it will be asked to the user using the getpass.getpass() function
+
+        Returns
+         - AuthorizationState. The called have to call `login` to continue the login process.
+
+        Raises:
+          - RuntimeError if the login failed
+
+        """
+        result = self._send_password(password)
+        self.authorization_state = self._wait_authorization_result(result)
+
+        return self.authorization_state